# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List the signed-in shopper's orders
      responses:
        "200":
          description: Shopper order history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Sign in required
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order and reserve inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order or unavailable inventory
        "401":
          description: Sign in required
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get one of the signed-in shopper's orders
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Shopper order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Sign in required
        "404":
          description: Order not found
  /vendors/{id}/orders/manage:
    get:
      operationId: listVendorOrders
      tags: [orders]
      summary: List orders containing a vendor's products
      parameters:
        - $ref: "#/components/parameters/VendorId"
      responses:
        "200":
          description: Vendor fulfillment orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendorOrder"
        "403":
          description: Vendor ownership required
  /vendors/{vendorId}/orders/{orderId}/items/{itemId}/status:
    patch:
      operationId: updateOrderItemStatus
      tags: [orders]
      summary: Update fulfillment status for a vendor's order item
      parameters:
        - $ref: "#/components/parameters/VendorOrderVendorId"
        - $ref: "#/components/parameters/VendorOrderId"
        - $ref: "#/components/parameters/OrderItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemStatusUpdate"
      responses:
        "200":
          description: Updated shopper order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          description: Vendor ownership required
        "404":
          description: Order item not found
components:
  parameters:
    VendorId:
      in: path
      name: id
      required: true
      schema:
        type: string
        minLength: 1
...
    OrderId:
      in: path
      name: id
      required: true
      schema:
        type: string
        minLength: 1
    OrderItemId:
      in: path
      name: itemId
      required: true
      schema:
        type: string
        minLength: 1
...
    VendorOrderVendorId:
      in: path
      name: vendorId
      required: true
      schema:
        type: string
        minLength: 1
    VendorOrderId:
      in: path
      name: orderId
      required: true
      schema:
        type: string
        minLength: 1
  schemas:
...
    OrderItemStatus:
      type: string
      enum: [pending, processing, shipped, delivered, cancelled]
    OrderItem:
      type: object
      required:
        [
          id,
          productId,
          vendorId,
          productName,
          vendorName,
          unitPrice,
          quantity,
          size,
          fulfillmentStatus
        ]
      properties:
        id: { type: string }
        productId: { type: string }
        vendorId: { type: string }
        productName: { type: string }
        vendorName: { type: string }
        imageUrl: { type: ["string", "null"] }
        unitPrice: { type: number, minimum: 0 }
        quantity: { type: integer, minimum: 1 }
        size: { type: string }
        fulfillmentStatus:
          $ref: "#/components/schemas/OrderItemStatus"
    OrderStatus:
      type: string
      enum: [pending, processing, shipped, delivered, cancelled]
    ShippingAddress:
      type: object
      required: [name, addressLine1, city, postalCode, country]
      properties:
        name: { type: string }
        addressLine1: { type: string }
        addressLine2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: ["string", "null"] }
        postalCode: { type: string }
        country: { type: string }
    OrderInputItem:
      type: object
      additionalProperties: false
      required: [productId, quantity, size]
      properties:
        productId: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1, maximum: 100 }
        size: { type: string, minLength: 1, maxLength: 30 }
    OrderInput:
      type: object
      additionalProperties: false
      required: [items, shippingAddress]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/OrderInputItem"
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
    Order:
      type: object
      required:
        [id, status, subtotal, shipping, total, shippingAddress, items, createdAt, updatedAt]
      properties:
        id: { type: string }
        status:
          $ref: "#/components/schemas/OrderStatus"
        subtotal: { type: number, minimum: 0 }
        shipping: { type: number, minimum: 0 }
        total: { type: number, minimum: 0 }
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    VendorOrder:
      allOf:
        - $ref: "#/components/schemas/Order"
    OrderItemStatusUpdate:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/OrderItemStatus"
...